Doom for the Gigabeat and some code reduction
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@12051 a1c6a512-1295-4272-9138-f99709370657

+++ apps/plugins/doom/g_game.c
@@ -2773,11 +2773,11 @@ boolean G_CheckDemoStatus (void)
       int fd=open(GAMEBASE "timedemo.txt",O_WRONLY | O_CREAT);
       fdprintf (fd,"Timed %d gametics in %d realtics = %d frames per second",
                (unsigned) gametic, realtics,
-               (unsigned) gametic * (double) TICRATE / realtics);
+               (unsigned) gametic * TICRATE/ realtics);
       close(fd);
       I_Error ("%d gametics in %d realtics",
                (unsigned) gametic,realtics,
-               (unsigned) gametic * (double) TICRATE / realtics);
+               (unsigned) gametic * TICRATE / realtics);
       return false;
    }
